requirements.txt
google-cloud-pubsub==0.40.0
pika==1.0.0
